fix https://github.com/waifuvault/WaifuVault/issues/344 and bump deps
