+++ package.json
@@ -27,58 +27,58 @@
     },
     "dependencies": {
         "@socket.io/redis-adapter": "^8.3.0",
-        "@tsed/ajv": "^8.22.0",
-        "@tsed/core": "^8.22.0",
-        "@tsed/di": "^8.22.0",
-        "@tsed/engines": "^8.22.0",
-        "@tsed/exceptions": "^8.22.0",
-        "@tsed/hooks": "^8.22.0",
-        "@tsed/ioredis": "^8.22.0",
-        "@tsed/json-mapper": "^8.22.0",
+        "@tsed/ajv": "^8.25.0",
+        "@tsed/core": "^8.25.0",
+        "@tsed/di": "^8.25.0",
+        "@tsed/engines": "^8.25.0",
+        "@tsed/exceptions": "^8.25.0",
+        "@tsed/hooks": "^8.25.0",
+        "@tsed/ioredis": "^8.25.0",
+        "@tsed/json-mapper": "^8.25.0",
         "@tsed/logger": "8.0.4",
         "@tsed/logger-file": "8.0.4",
         "@tsed/logger-std": "^8.0.4",
-        "@tsed/passport": "^8.22.0",
-        "@tsed/platform-cache": "^8.22.0",
-        "@tsed/platform-exceptions": "^8.22.0",
-        "@tsed/platform-express": "^8.22.0",
-        "@tsed/platform-http": "^8.22.0",
-        "@tsed/platform-log-middleware": "^8.22.0",
-        "@tsed/platform-log-request": "^8.22.0",
-        "@tsed/platform-middlewares": "^8.22.0",
-        "@tsed/platform-multer": "^8.22.0",
-        "@tsed/platform-params": "^8.22.0",
-        "@tsed/platform-response-filter": "^8.22.0",
-        "@tsed/platform-views": "^8.22.0",
-        "@tsed/scalar": "^8.22.0",
-        "@tsed/schema": "^8.22.0",
-        "@tsed/socketio": "^8.22.0",
-        "@tsed/swagger": "^8.22.0",
+        "@tsed/passport": "^8.25.0",
+        "@tsed/platform-cache": "^8.25.0",
+        "@tsed/platform-exceptions": "^8.25.0",
+        "@tsed/platform-express": "^8.25.0",
+        "@tsed/platform-http": "^8.25.0",
+        "@tsed/platform-log-middleware": "^8.25.0",
+        "@tsed/platform-log-request": "^8.25.0",
+        "@tsed/platform-middlewares": "^8.25.0",
+        "@tsed/platform-multer": "^8.25.0",
+        "@tsed/platform-params": "^8.25.0",
+        "@tsed/platform-response-filter": "^8.25.0",
+        "@tsed/platform-views": "^8.25.0",
+        "@tsed/scalar": "^8.25.0",
+        "@tsed/schema": "^8.25.0",
+        "@tsed/socketio": "^8.25.0",
+        "@tsed/swagger": "^8.25.0",
         "ajv": "^8.17.1",
         "archiver": "^7.0.1",
         "argon2": "^0.44.0",
-        "better-sqlite3": "^12.6.0",
+        "better-sqlite3": "^12.6.2",
         "body-parser": "^2.2.2",
         "builder-pattern": "^2.2.0",
         "compression": "^1.8.1",
         "connect-redis": "^8.1.0",
         "cookie-parser": "^1.4.7",
-        "cors": "^2.8.5",
-        "cronstrue": "^3.9.0",
+        "cors": "^2.8.6",
+        "cronstrue": "^3.11.0",
         "cross-env": "^10.1.0",
-        "dotenv": "^17.2.3",
+        "dotenv": "^17.2.4",
         "dotenv-expand": "^12.0.3",
         "dotenv-flow": "^4.1.0",
         "express": "^5.2.1",
         "express-rate-limit": "^8.2.1",
-        "express-session": "^1.18.2",
+        "express-session": "^1.19.0",
         "ffmpeg-ffprobe-static": "^6.1.2-rc.1",
         "file-type": "^21.3.0",
         "fluent-ffmpeg": "^2.1.3",
         "helmet": "^8.1.0",
         "http-status-codes": "^2.3.0",
         "ioredis": "^5.8.1",
-        "is-localhost-ip": "^2.0.0",
+        "is-localhost-ip": "^3.0.1",
         "method-override": "^3.0.0",
         "mime": "^4.1.0",
         "node-schedule": "^2.1.1",
@@ -86,7 +86,7 @@
         "passport-custom": "^1.1.1",
         "passport-http": "^0.3.0",
         "passport-local": "^1.0.0",
-        "pg": "^8.16.3",
+        "pg": "^8.18.0",
         "punycode": "^2.3.1",
         "redis": "^4.7.1",
         "socket.io": "^4.8.3",
@@ -99,8 +99,8 @@
     },
     "devDependencies": {
         "@eslint/js": "^9.39.2",
-        "@swc/cli": "0.7.9",
-        "@swc/core": "^1.15.8",
+        "@swc/cli": "0.8.0",
+        "@swc/core": "^1.15.11",
         "@tsed/barrels": "^6.6.3",
         "@tsed/cli-plugin-passport": "6.6.3",
         "@types/archiver": "^7.0.0",
@@ -114,7 +114,7 @@
         "@types/jsdom": "^27.0.0",
         "@types/method-override": "^3.0.0",
         "@types/multer": "^2.0.0",
-        "@types/node": "^25.0.6",
+        "@types/node": "^25.2.1",
         "@types/node-schedule": "^2.1.8",
         "@types/passport": "^1.0.17",
         "@types/passport-http": "^0.3.11",
@@ -124,14 +124,14 @@
         "husky": "^9.1.7",
         "lint-staged": "^16.2.7",
         "nodemon": "^3.1.11",
-        "prettier": "^3.7.4",
+        "prettier": "^3.8.1",
         "ts-node": "^10.9.2",
         "ts-node-dev": "^2.0.0",
         "tsc-files": "^1.1.4",
         "tsconfig-paths": "^4.2.0",
         "tslib": "^2.8.1",
         "typescript": "5.9.3",
-        "typescript-eslint": "^8.52.0"
+        "typescript-eslint": "^8.54.0"
     },
     "main": "index.js",
     "keywords": [],
